Перенос фильтра в отдельный модуль
