+++ js/main.js
@@ -3,6 +3,6 @@ import './user-forms.js';
 import {map} from './map.js';
 map;
 import './server-requests.js';
-
+import './map-filters.js';
